chore: relocate journal pages, components, and helpers
Split the journal module out of Pages/Journal into one folder per page
and per shared component, and move pure helpers into utils/ and data/.
Styles still share Journal.css; that split lands in a follow-up commit.
