+++ frontend/src/Pages/JournalHistory/JournalHistory.jsx
@@ -8,8 +8,8 @@ import { useDemoMode } from '../../Context/DemoModeContext.jsx';
 import { apiFetch } from '../../api/client.js';
 import { DEMO_ENTRIES } from '../../data/demoJournal.js';
 import { emitToast } from '../../lib/toastBus.js';
-import JournalSummaryBanner from './JournalSummaryBanner.jsx';
-import './Journal.css';
+import JournalSummaryBanner from '../../Components/JournalSummaryBanner/JournalSummaryBanner.jsx';
+import '../Journal/Journal.css';
 
 const EXCERPT_WORD_COUNT = 40;
