+++ frontend/src/Pages/JournalSummary/JournalSummary.test.jsx
