+++ frontend/src/data/summaryLoadingQuotes.js
