+++ frontend/src/utils/journalImage.js
