+++ frontend/src/utils/journalImage.test.js
