+++ frontend/src/utils/journalTopicsFade.js
