+++ frontend/src/utils/journalTopicsFade.test.js
